requirements.txt
aenum==3.1.11
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==2.1.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
frozenlist==1.3.3
gremlinpython==3.6.2
idna==3.4
isodate==0.6.1
kiwisolver==1.4.4
matplotlib==3.7.0
multidict==6.0.4
neo4j==5.3.0
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
seaborn==0.12.2
six==1.16.0
yarl==1.8.2
